#33 wip split utils directory
