+++ src/components/ChipSelect/ChipSelect.tsx
@@ -3,8 +3,8 @@ import {
   type FunctionComponent,
   type KeyboardEvent,
 } from 'react';
-import { classNames } from '../../utils/common';
 import { Chip } from '../Chip';
+import { classNames } from '../../utils';
 import { useChipSelect } from './hooks/useChipSelect';
 import { useClickOutside } from './hooks/useClickOutside';
 import { useScrollToFocused } from './hooks/useScrollToFocused';
